Move helper methods from parallel runner to task
