+++ pyproject.toml
@@ -170,6 +170,7 @@ layers = [
 ]
 ignore_imports = [
     "kedro.runner.parallel_runner -> kedro.framework.project",
+    "kedro.runner.task -> kedro.framework.project",
     "kedro.framework.hooks.specs -> kedro.framework.context"
 ]
